dm vdo test: fix type of targetBlockedThreadCount in DataVIOPool_t1
checkBlockedThreadCount() in mutexUtils.c dereferences its context
argument as uint32_t *, but targetBlockedThreadCount was declared as
uint8_t. ASAN reports this as a global-buffer-overflow: a 4-byte read
of a 1-byte global.

+++ src/c++/vdo/tests/DataVIOPool_t1.c
@@ -55,7 +55,7 @@ static struct data_vio *blocked[REQUEST_COUNT + DATA_VIO_COUNT];
 static uint8_t          blockedCount;
 static uint8_t          nextLBNExpected;
 static struct thread   *threads[REQUEST_COUNT];
-static uint8_t          targetBlockedThreadCount;
+static uint32_t         targetBlockedThreadCount;
 
 /**
  * Test-specific initialization.
